feat(trusty-mpm): port 11 mpm-* guidance skills into bundled assets (closes #770 Phase 1)
STEP 0 decision: trusty-mpm deployer expects flat *.md files (not SKILL.md
directories), so all 11 skills are written as flat files under
crates/trusty-mpm/src/assets/skills/ matching the existing example-skill.md
pattern. No deployer changes required.

Skills ported (sourced from .claude/skills/, adapted to trusty-mpm conventions):
  mpm-delegation-patterns       — PM agent type selection & evidence chains
  mpm-verification-protocols    — QA gate & evidence requirements
  mpm-git-file-tracking         — blocking git-tracking gate protocol
  mpm-pr-workflow               — feature-branch + PR creation workflow
  mpm-ticketing-integration     — ticket state transitions & comment trails
  mpm-circuit-breaker-enforcement — full circuit-breaker catalogue
  mpm-bug-reporting             — routing tree & delegation template
  mpm-session-management        — context-limit session continuity guide
  mpm-session-pause             — /mpm-session-pause slash command (PM instructions)
  mpm-session-resume            — /mpm-session-resume slash command (PM instructions)
  mpm-tool-usage-guide          — tool usage patterns for PM agents

Adaptations applied to every skill:
  ~/.claude-mpm/  →  ~/.trusty-mpm/
  claude-mpm CLI  →  tm / trusty-mpm
  config.yaml     →  config.toml
  bug repo        →  bobmatnyc/trusty-tools
  session dir     →  .trusty-mpm/sessions/
  Python-MPM implementation code removed from pause/resume skills

bundle.rs refactored to stay under the 500-line cap:
  constants extracted to bundle_skills.rs (106 lines)
  BundledArtifact / InstallPolicy / ALL moved to bundle_all.rs (338 lines)
  bundle.rs reduced to constants + pub use facade (191 lines)
  ALL.len() grows from 46 → 57 (+11 guidance skills)

Tests updated:
  bundle_table_is_complete: 46 → 57
  constants_are_non_empty: +11 skill const assertions
  phase1_guidance_skills_are_in_bundle: new — checks all 11 paths in ALL
  phase1_guidance_skills_have_frontmatter: new — checks frontmatter + no unadapted claude-mpm refs
  install_then_deploy_deploys_skills: assert 12 skill files deployed (1 placeholder + 11 guidance)

Quality bar: cargo check ✓  clippy -D warnings ✓  cargo test ✓  cargo fmt --check ✓  check_line_cap.sh ✓

Part of #770 — Phase 1: core mpm-* guidance skills (sourced from claude-mpm-skills)
